refactoring naming csv importer
