+++ backend/csvimporter/src/main/java/com/simra/app/csvimporter/repository/CSVFileRepository.java
@@ -1,4 +1,4 @@
-package com.simra.app.csvimporter.controller;
+package com.simra.app.csvimporter.repository;
 
 import com.simra.app.csvimporter.model.CSVFile;
 import org.springframework.data.mongodb.repository.MongoRepository;
