+++ backend/csvimporter/src/main/java/com/simra/app/csvimporter/repository/LegRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.simra.app.csvimporter.controller;
+package com.simra.app.csvimporter.repository;
 
 import com.simra.app.csvimporter.model.LegEntity;
