+++ backend/csvimporter/src/main/java/com/simra/app/csvimporter/service/RideParserThreaded.java
@@ -4,8 +4,8 @@ import com.mongodb.client.model.geojson.LineString;
 import com.mongodb.client.model.geojson.Position;
 import com.opencsv.bean.ColumnPositionMappingStrategy;
 import com.opencsv.bean.CsvToBeanBuilder;
-import com.simra.app.csvimporter.controller.IncidentRepository;
-import com.simra.app.csvimporter.controller.RideRepository;
+import com.simra.app.csvimporter.repository.IncidentRepository;
+import com.simra.app.csvimporter.repository.RideRepository;
 import com.simra.app.csvimporter.filter.MapMatchingService;
 import com.simra.app.csvimporter.filter.RideSmoother;
 import com.simra.app.csvimporter.model.RideCSV;
